Remove dead code and fix tests
Dead code removed according to golangci-lint report. Test errors arose
from the new wait-to-speak behaviour of the hub, so the test echo
handler mocking sshd was updated to send a greeting upon connection, and
the tests to check for this before carrying to wait for the
echo. Shellhost and shellclient tests now passing again. Also, an over
zealous test failure condition in pkg/counter was causing a test fail
whenever a non-mutexed count failed to encouter errors - this is not a
test fail, just an indication of the statistical nature of the issue
without mutexing.