Relax timing in test pkg/agg
Increase timeout from 3ms to 10ms in TestStreamWithRuleChange because it
failed once in a blue moon for me just now, with subsequent testing
showing it was failing by 0.5ms. If this newer more-relaxed test starts to fail
consistently, then we have a performance degradation somewhere, and
benchmarking/profiling will be required to evaluate.


Former-commit-id: 0cf75731f5c9d7a3bf999602df5cc40e34f184ab