Add pkg/bookingstore
Package bookingstore is an upgrade to pkg/limit, which retains the
session limiting functionality, but also stores a pointer
to the activity, so that current user sessions can be requested (TODO). The
confirm function takes the pointer as an argument. A cancel function is
return, so that the booking can be explicitly cancelled, quickly.


Former-commit-id: 722b09b04d3aeb70b9fe3e8d091cf0ab2427bea3