Add pkg/bookingstore
Package bookingstore is an upgrade to pkg/limit, which retains the
session limiting functionality, but also stores a pointer
to the activity, so that current user sessions can be requested (TODO). The
confirm function takes the pointer as an argument. A cancel function is
return, so that the booking can be explicitly cancelled, quickly.